+++ client/src/components/close-button.stories.tsx
