+++ client/src/components/close-button.tsx
