+++ client/src/components/navbarlogoutview.tsx
@@ -2,17 +2,19 @@ import { Button } from "./button";
 
 //To clicked and open the login?signup modal
 interface NavBarLoggedOutViewProps {
-  // onSignUpClicked: () => void;
+  onSignUpClicked: () => void;
   onLoginClicked: () => void;
 }
 
 export const NavBarLogOutView = ({
-  // onSignUpClicked,
+  onSignUpClicked,
   onLoginClicked,
 }: NavBarLoggedOutViewProps) => {
   return (
     <div className="flex flex-col lg:flex-row lg:ml-auto gap-4">
-      <Button variant={"primary"}>Sign Up</Button>
+      <Button variant={"primary"} onClick={onSignUpClicked}>
+        Sign Up
+      </Button>
       <Button variant={"primary"} onClick={onLoginClicked}>
         Log In
       </Button>
