+++ client/src/components/signupmodal.tsx
@@ -3,7 +3,8 @@ import { Input } from "./input";
 import { Modal } from "./modal";
 import { User } from "../models/user";
 import { useState } from "react";
-import { LogInCredentials } from "../api/users_api";
+import { SignUpCredentials } from "../api/users_api";
+import * as UsersApi from "../api/users_api";
 
 interface SignUpModalProps {
   onDismiss: () => void;
@@ -30,10 +31,25 @@ export const SignUpModal = ({
       [event.target.name]: event.target.value,
     }));
   };
-
+  async function handleSubmit(credentials: SignUpCredentials) {
+    try {
+      const user = await UsersApi.signUp(credentials);
+      console.log(user);
+      onSignUpSuccessful(user);
+    } catch (error) {
+      alert(error);
+      console.log(error);
+    }
+  }
+  const handleSubmitLog = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    handleSubmit(data);
+    console.log(data);
+    setData(initialData);
+  };
   return (
     <Modal isOpen={true} title={"Sign Up"}>
-      <form>
+      <form onSubmit={handleSubmitLog}>
         <Input
           label="UserName"
           type="text"
