+++ client/src/pages/LogoutPage.tsx
@@ -1,0 +1,7 @@
+export const LogoutPage = () => {
+  return (
+    <div className="px-6 w-full h-full m-auto min-h-screen">
+      <p>Please login to see your reports</p>
+    </div>
+  );
+};
